Create coursera 2week 2 gladlibmap
Assignment 3: Maps Version of GladLibs
Start with your GladLibs program you completed earlier in this lesson. Make a copy of it and call
it GladLibMap.java. Now modify this program to use one HashMap that maps word types to
ArrayList of possible words to select. Your program should still work for the additional categories
verbs and fruits and should not use duplicate words from a category. Specifically, you should
make the following adjustments to this program:
● Replace the ArrayLists for adjectiveList,nounList,colorList, countryList,nameList,
animalList, timeList, verbList,and fruitListwith one HashMap myMapthat maps a
String representing a category to an ArrayList of words in that category. Caution: Don’t
replace the ArrayList representing the words that have already been used!
● Create the new HashMap in the constructors.
● Modify the method initializeFromSourceto create an Array of categories and then
iterate over this Array. For each category, read in the words from the associated file,
create an ArrayList of the words (using the method readIt), and put the category and
ArrayList into the HashMap.
● Modify the method getSubstituteto replace all the if statements that use category labels
with one call to randomFromthat passes the appropriate ArrayList from myMap.
● Run your program to make sure it works.
● Write a new method named totalWordsInMapwith no parameters. This method returns
the total number of words in all the ArrayLists in the HashMap. After printing the GladLib,
call this method and print out the total number of words that were possible to pick from.
● Write a new method named totalWordsConsideredwith no parameters. This method
returns the total number of words in the ArrayLists of the categories that were used for a
particular GladLib. If only noun, color, and adjective were the categories used in a
GladLib, then only calculate the sum of all the words in those three ArrayLists. Hint: You
will need to keep track of the categories used in solving the GladLib, then compute this
total.